🐛 Move mockShopifyAnalytics test helper out of src to fix package build
