+++ packages/browser-rum-shopify/src/domain/shopifyPlugin.spec.ts
@@ -2,7 +2,7 @@ import { TIMEOUT_ERROR_MESSAGE } from '@datadog/browser-core'
 import { mockClock, replaceMockableWithSpy } from '@datadog/browser-core/test'
 import type { RumInitConfiguration, RumPublicApi } from '@datadog/browser-rum-core'
 import { patchSandboxedIframeApis } from '../boot/patchSandboxedIframeApis'
-import { createFakeAnalytics, pageViewedEvent } from '../test/mockShopifyAnalytics'
+import { createFakeAnalytics, pageViewedEvent } from '../../test/mockShopifyAnalytics'
 import type { ShopifyAnalyticsApi } from './shopifyAnalytics'
 import { initShopifyBindings } from './shopifyBindings'
 import { shopifyPlugin } from './shopifyPlugin'
